Use location_id stored in db
By creating a new constructor that takes it from an argument. It's kind of weird to concatenate together an id from its part when I have the actual ID stored. This didn't work when I was trying out the notify lambda and used bad test data (id not corresponding to street/district)