Sort list of stops, and remove any duplicates
My initial hope was that I wouldn't have a need to remove duplicates. Basically because with the byLocationId index I use location_id as hash key, so my thought was that only unique items would exist. Turns out however that DynamoDb in some way includes the range key of the base index anyway, and still list multiple items with the same hash key